// src/types/trip.ts
export enum TripStatus {
  DRAFT = 'draft',
  PUBLISHED = 'published',
  FULL = 'full',
  IN_PROGRESS = 'in_progress',
  COMPLETED = 'completed',
  CANCELLED = 'cancelled'
}

export enum TripType {
  DOMESTIC = 'domestic',
  INTERNATIONAL = 'international'
}

export enum TransportMode {
  CAR = 'car',
  TRAIN = 'train',
  PLANE = 'plane',
  BUS = 'bus'
}